a little optimize

+++ eeprom_writer.c
@@ -13,14 +13,14 @@ int main () {
 	TFT_on();
 	print_str(10,22,1,White,Black,"Writing EEPROM...");
 	for(i=0; i < SDrive_eep_bin_len; i++) {
-		Draw_H_Bar(10,229,50,10,(int) (i/(SDrive_eep_bin_len/219.0)),Yellow,Blue,Grey,1);
+		Draw_H_Bar(10,229,50,10,(i/(SDrive_eep_bin_len/219.0)),Yellow,Blue,Grey,1);
 		//eeprom_write_byte((unsigned char*)i, pgm_read_byte(&SDrive_eep_bin[i]));
 		eeprom_update_byte((unsigned char*)i, SDrive_eep_bin[i]);
 		//_delay_ms(10);
 	}
 	print_str(10,122,1,White,Black,"Verifying EEPROM...");
 	for(i=0; i < SDrive_eep_bin_len; i++) {
-		Draw_H_Bar(10,229,150,10,(int) (i/(SDrive_eep_bin_len/219.0)),Yellow,Blue,Grey,1);
+		Draw_H_Bar(10,229,150,10,(i/(SDrive_eep_bin_len/219.0)),Yellow,Blue,Grey,1);
 		if (eeprom_read_byte((unsigned char*)i) != SDrive_eep_bin[i]) {
 			print_str(10,180,1,Red,Black,"Verify Error!");
 			return(1);
